apennd issue solved.
Put queue q[] creation in test case. This gives me a new queue for every test case. I think I need to free that previous queue first... I need to revisit this. Also This code seemed to work well in msvc but line 181 was causing issues in repl. sizeOf[customers) is giving me 8 in repl and giving me the proper number in msvc... I may need to create m in main and pass by reference in to createCustomers to reuse in main. This is for properly queing customers.