output looks good but not complete
In this version I was able to get the names to uppercase and get some of the dequeuing in. In this version of the code I am stopping 1 short with a segmentation fault on line 216. On the 5th iteration my nextCustomer is showing 5 when it should be 11 so I am possibly not dequeuing correctly. I also need to work out the logic regarding the served time. I am getting the first one correct but I am so far unable to carry that time over to the next iteration. I also need to add in the numOfSmoothie caveat.

Input:
1
5
10 1 imran 12
12 6 adam 8
13 1 mehmed 40
22 6 chris 39
100000 12 orhan 53

output:
At time 100, IMRAN left the counter from line 1
At time 82, ADAM left the counter from line 6
At time 243, MEHMED left the counter from line 1
At time 247, CHRIS left the counter from line 6